+++ src/main.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from '@/App';
 import '@/index.css';
-import { CartProvider } from '@/hooks/useCart';
 import { AuthProvider } from '@/contexts/SupabaseAuthContext';
 
 import { HelmetProvider } from 'react-helmet-async';
@@ -11,9 +10,7 @@ ReactDOM.createRoot(document.getElementById('root')).render(
     <React.StrictMode>
         <HelmetProvider>
             <AuthProvider>
-                <CartProvider>
-                    <App />
-                </CartProvider>
+                <App />
             </AuthProvider>
         </HelmetProvider>
     </React.StrictMode>